- Made tests compatible with ZODB 3.9. - added coverage analysis tools to buildout.cfg - removed *-configure.zcml slug

+++ src/zope/file/testing.py
@@ -26,12 +26,14 @@ from ZODB.blob import BlobStorage
 from zope.testing import doctest
 import zope.app.testing.functional
 from zope.app.component.hooks import setSite
+import ZODB.interfaces
 
 here = os.path.dirname(os.path.realpath(__file__))
 
 class FunctionalBlobTestSetup(zope.app.testing.functional.FunctionalTestSetup):
 
     temp_dir_name = None
+    direct_blob_support = False
 
     def setUp(self):
         """Prepares for a functional test case."""
@@ -39,10 +41,14 @@ class FunctionalBlobTestSetup(zope.app.testing.functional.FunctionalTestSetup):
         transaction.abort()
         self.db.close()
         storage = DemoStorage("Demo Storage", self.base_storage)
-        # make a dir
-        temp_dir_name = self.temp_dir_name = tempfile.mkdtemp()
-        # wrap storage with BlobStorage
-        storage = BlobStorage(temp_dir_name, storage)
+        if ZODB.interfaces.IBlobStorage.providedBy(storage):
+            # at least ZODB 3.9
+            self.direct_blob_support = True
+        else:
+            # make a dir
+            temp_dir_name = self.temp_dir_name = tempfile.mkdtemp()
+            # wrap storage with BlobStorage
+            storage = BlobStorage(temp_dir_name, storage)
         self.db = self.app.db = DB(storage)
         self.connection = None
 
@@ -53,8 +59,8 @@ class FunctionalBlobTestSetup(zope.app.testing.functional.FunctionalTestSetup):
             self.connection.close()
             self.connection = None
         self.db.close()
-        # del dir named '__blob_test__%s' % self.name
-        if self.temp_dir_name is not None:
+        if not self.direct_blob_support and self.temp_dir_name is not None:
+            # del dir named '__blob_test__%s' % self.name
             shutil.rmtree(self.temp_dir_name, True)
             self.temp_dir_name = None
         setSite(None)
